# actions/openapi.yaml
openapi: 3.1.0
info:
  title: Trello Actions Proxy
  version: 0.1.0
servers:
  - url: https://trello-actions-proxy-kl2lm6izya-an.a.run.app
paths:
  /v1/ping:
    get:
      operationId: ping
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PingResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/trello/boards:
    get:
      operationId: listBoards
      summary: List Trello boards
      responses:
        "200":
          description: List of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Board"
        "401":
          description: Trello authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Trello rate limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Trello error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/trello/boards/{boardId}/lists:
    get:
      operationId: listLists
      summary: List lists for a Trello board
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/List"
        "403":
          description: Board is not allowlisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Trello authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Trello rate limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Trello error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/trello/boards/{boardId}/cards:
    get:
      operationId: listBoardCards
      summary: List cards for a Trello board
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
        "403":
          description: Board is not allowlisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Trello authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Trello rate limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Trello error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/trello/lists/{listId}/cards:
    get:
      operationId: listListCards
      summary: List cards for a Trello list
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
        "403":
          description: Board is not allowlisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Trello authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Trello rate limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Trello error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/trello/cards/{cardId}:
    get:
      operationId: getCard
      summary: Get Trello card details
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "403":
          description: Board is not allowlisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Trello authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Trello rate limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Trello error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/trello/write/preview:
    post:
      operationId: previewWrite
      summary: Preview write operations
      parameters:
        - name: X-Internal-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WritePreviewRequest"
      responses:
        "200":
          description: Preview response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WritePreviewResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/trello/write/commit:
    post:
      operationId: commitWrite
      summary: Commit write operations
      parameters:
        - name: X-Internal-Token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteCommitRequest"
      responses:
        "200":
          description: Commit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WriteCommitResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    PingResponse:
      type: object
      properties:
        ok:
          type: boolean
        service:
          type: string
        version:
          type: string
      required:
        - ok
        - service
        - version
    Board:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        closed:
          type: boolean
      required:
        - id
        - name
        - url
        - closed
    List:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        closed:
          type: boolean
      required:
        - id
        - name
        - closed
    Card:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        desc:
          type: string
        url:
          type: string
        idList:
          type: string
        idBoard:
          type: string
        due:
          type:
            - string
            - "null"
        dueComplete:
          type: boolean
        closed:
          type: boolean
        labels:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              color:
                type:
                  - string
                  - "null"
            required:
              - id
              - name
      required:
        - id
        - name
        - desc
        - url
        - idList
        - idBoard
        - dueComplete
        - closed
        - labels
    Comment:
      type: object
      properties:
        id:
          type: string
        cardId:
          type: string
        text:
          type: string
      required:
        - id
        - cardId
        - text
    WriteAction:
      type: string
      enum:
        - createCard
        - addComment
        - updateCard
    CreateCardPayload:
      type: object
      properties:
        listId:
          type: string
        name:
          type: string
        desc:
          type: string
        due:
          type:
            - string
            - "null"
        dueComplete:
          type: boolean
        labelIds:
          type: array
          items:
            type: string
      required:
        - listId
        - name
    AddCommentPayload:
      type: object
      properties:
        cardId:
          type: string
        text:
          type: string
      required:
        - cardId
        - text
    UpdateCardPayload:
      type: object
      properties:
        cardId:
          type: string
        name:
          type: string
        desc:
          type: string
        due:
          type:
            - string
            - "null"
        dueComplete:
          type: boolean
        closed:
          type: boolean
        listId:
          type: string
        labelIds:
          type: array
          items:
            type: string
      required:
        - cardId
    WritePreviewRequest:
      type: object
      properties:
        action:
          $ref: "#/components/schemas/WriteAction"
        payload:
          oneOf:
            - $ref: "#/components/schemas/CreateCardPayload"
            - $ref: "#/components/schemas/AddCommentPayload"
            - $ref: "#/components/schemas/UpdateCardPayload"
      required:
        - action
        - payload
    WritePreviewResponse:
      type: object
      properties:
        preview:
          type: object
          properties:
            action:
              $ref: "#/components/schemas/WriteAction"
            payload:
              type: object
          required:
            - action
            - payload
        commitToken:
          type: string
        expiresAt:
          type: string
      required:
        - preview
        - commitToken
        - expiresAt
    WriteCommitRequest:
      type: object
      properties:
        commitToken:
          type: string
      required:
        - commitToken
    WriteCommitResponse:
      type: object
      properties:
        action:
          $ref: "#/components/schemas/WriteAction"
        result:
          oneOf:
            - $ref: "#/components/schemas/Card"
            - $ref: "#/components/schemas/Comment"
      required:
        - action
        - result
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
          required:
            - code
            - message
      required:
        - error